Implemented Delete HabitLog Function
#5
