--- src/HabitTracker.UI/Pages/Delete.cshtml
@@ -1,31 +1,0 @@
-﻿@page
-@model DeleteModel
-@{
-
-}
-
-<h3>Are you sure you want to delete this record?</h3>
-
-<div>
-    <dl class="row">
-        <dt class="col-sm-2">
-            @Html.DisplayNameFor(x => x.DrinkingWater.Date)
-        </dt>
-        <dd class="col-sm-10">
-            @Html.DisplayFor(x => x.DrinkingWater.Date)
-        </dd>
-        <dt class="col-sm-2">
-            @Html.DisplayNameFor(x => x.DrinkingWater.Quantity)
-        </dt>
-        <dd class="col-sm-10">
-            @Html.DisplayFor(x => x.DrinkingWater.Quantity)
-        </dd>
-    </dl>
-
-    <form method="post">
-        <input type="hidden" asp-for="DrinkingWater.Id" />
-        <input type="submit" value="Delete" class="btn btn-danger" />
-        <br />
-        <a asp-page="./Index">Back to List</a>
-    </form>
-</div>
