+++ src/HabitTracker.UI/Pages/DeleteHabitLog.cshtml
@@ -1,0 +1,39 @@
+﻿@page
+@model HabitTracker.WebUI.Pages.DeleteHabitLogModel
+@{
+    ViewData["Title"] = "Delete Habit Log";
+}
+
+<h2>Delete TODO HabitName Logs</h2>
+<hr />
+<h3>Are you sure you want to delete this record?</h3>
+<div>
+    <dl class="row">
+        <dt class="col-sm-2">
+            @Html.DisplayNameFor(x => x.HabitLog!.Date)
+        </dt>
+        <dd class="col-sm-10">
+            @Html.DisplayFor(x => x.HabitLog!.Date)
+        </dd>
+        <dt class="col-sm-2">
+            @Html.DisplayNameFor(x => x.HabitLog!.Quantity)
+        </dt>
+        <dd class="col-sm-10">
+            @Html.DisplayFor(x => x.HabitLog!.Quantity)
+        </dd>
+    </dl>
+
+    <form method="post">
+        <input type="hidden" asp-for="HabitLog!.Id" />
+        <input type="hidden" asp-for="HabitLog!.HabitId" />
+        <input type="submit" value="Delete" class="btn btn-danger" />
+        <br />
+        <a asp-page="./Index">Back to List</a>
+    </form>
+</div>
+
+@section Scripts {
+    @{
+        await Html.RenderPartialAsync("_ValidationScriptsPartial");
+    }
+}
